+++ apps/mission/e2e/cleanupTabs.ts
@@ -1,0 +1,43 @@
+import { pageListener } from "@evmosapps/test-utils";
+import { E } from "helpers";
+import { BrowserContext } from "playwright";
+
+export const cleanupTabs = async (context: BrowserContext) => {
+  const pages = context.pages();
+  for (const page of pages) {
+    if (page.url().startsWith("chrome-extension://")) {
+      await page.close();
+    }
+  }
+};
+export const getMMPopup = async (context: BrowserContext) => {
+  const pages = context.pages();
+  return (
+    pages.find((page) => page.url().startsWith("chrome-extension://")) ||
+    (await context.waitForEvent("page"))
+  );
+};
+export const connectSwitchAndSignPubkey = async (
+  context: BrowserContext,
+  trigger: () => Promise<void>
+) => {
+  const approveAllPopup = pageListener(context);
+
+  await trigger();
+  await approveAllPopup.load;
+  let popupPage = approveAllPopup.page;
+
+  await popupPage.getByRole("button", { name: /Next/i }).click();
+  await popupPage.getByRole("button", { name: /Connect/i }).click();
+
+  while (true) {
+    const [err] = await E.try(() =>
+      popupPage.getByRole("button", { name: /Sign/i }).click()
+    );
+    if (err?.message.includes("Page closed")) {
+      popupPage = await getMMPopup(context);
+      continue;
+    }
+    break;
+  }
+};