+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+- (fix) fse-837 | apps/vesting 1.1.4 | Fix error in account details page for vesting
 - (fix) fse-836 | apps/vesting 1.1.3 | Fix error on creating vesting account
 - (fix) fse-834 | apps/assets 1.0.38 | Fix error generating txns appear twice in pay modal
 - (fix) fse-800 | apps/assets 1.0.37 apps/governance 1.0.28 apps/staking 1.1.2 apps/vesting 1.1.2 packages/evmos-wallet 1.0.23 | Suggest changing network to evmos (MetaMask)